+++ sdk/lib/opentelemetry/sdk/internal.rb
@@ -44,12 +44,12 @@ module OpenTelemetry
         valid_simple_value?(value) || valid_array_value?(value)
       end
 
-      def valid_attributes?(owner, kind, attrs)
+      def valid_attributes?(owner, kind, attrs, length_limit: nil)
         attrs.nil? || attrs.all? do |k, v|
           if !valid_key?(k)
             OpenTelemetry.handle_error(message: "invalid #{kind} attribute key type #{k.class} on span '#{owner}'")
             false
-          elsif !valid_value?(v)
+          elsif !valid_value?(v) || (!length_limit.nil? && v.is_a?(String) && v.length > length_limit)
             OpenTelemetry.handle_error(message: "invalid #{kind} attribute value type #{v.class} for key '#{k}' on span '#{owner}'")
             false
           else